--- src/main/java/uk/gov/hmcts/reform/civil/model/caseprogression/LocationEntry.java
@@ -1,18 +1,0 @@
-package uk.gov.hmcts.reform.civil.model.caseprogression;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-
-@Data
-@Builder(toBuilder = true)
-@NoArgsConstructor
-@AllArgsConstructor
-public class LocationEntry {
-    private String type;
-    private String value;
-    private Map<String, Object> valueInfo;
-}
