+++ src/main/java/uk/gov/hmcts/reform/civil/model/caseprogression/SecondaryListingLocations.java
@@ -6,13 +6,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Map;
-
 @Data
 @Builder(toBuilder = true)
 @NoArgsConstructor
 @AllArgsConstructor
 public class SecondaryListingLocations  {
+
     @JsonProperty("PrelistingAdminLocation")
     private SecondaryLocationModel prelistingAdminLocation;
     @JsonProperty("CMCListingLocation")
