+++ src/main/java/uk/gov/hmcts/reform/civil/service/camunda/CamundaRuntimeApi.java
@@ -2,7 +2,11 @@ package uk.gov.hmcts.reform.civil.service.camunda;
 
 import org.camunda.community.rest.client.model.VariableValueDto;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 import java.util.HashMap;
 import java.util.List;
